+++ apps/chronicles/App.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useMemo } from "react";
+import Navigation from "./navigation";
 import { StatusBar } from "expo-status-bar";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 import { useEDS, EDSProvider } from "@equinor/mad-components";
 import { Envelope, addTelemetryInitializer, ErrorBoundary } from "@equinor/mad-core";
 import useCachedResources from "./hooks/useCachedResources";
 import useColorScheme from "./hooks/useColorScheme";
-import Navigation from "./navigation";
 import { useWindowDimensions } from "react-native";
 import * as APP from "./app.json";