+++ packages/components/src/index.ts
@@ -21,6 +21,7 @@ export * from "./components/Environment";
 export * from "./components/ErrorBoundary";
 export * from "./components/SelectionControls";
 export * from "./components/Chip";
+export * from "./components/Tabs";
 
 export * from "./styling/EDSStyleSheet";
 export * from "./styling/types";