+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @equinor/mad-auth
 
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [a5607df]
+-   Updated dependencies [0d6c310]
+    -   @equinor/mad-components@0.8.6
+
+## 0.1.3
+
+### Patch Changes
+
+-   b025c9b: Implemented Create Incident Screen
+-   Updated dependencies [b025c9b]
+    -   @equinor/mad-components@0.8.5
+
 ## 0.1.2
 
 ### Patch Changes