+++ packages/components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @equinor/mad-components
 
+## 0.8.6
+
+### Patch Changes
+
+-   a5607df: Created `Tabs` component!
+-   0d6c310: update color scheme for feedback and added deprecation
+
+## 0.8.5
+
+### Patch Changes
+
+-   b025c9b: Implemented Create Incident Screen
+
 ## 0.8.4
 
 ### Patch Changes